Added gameId's of blueprints' origins
Each function has an additional comment that contains the `gameId` of its returning blueprint's origin replay.

This is also a fix on the formatting of the blueprint file. This also serves as a reminder to new pull request starters to set your files' line break types to "Unix (LF)" before uploading to GitHub, or you will end up with `^M`s in place of line breaks.